+++ packages/apollo-server-koa/src/__tests__/koaApollo.test.ts
@@ -8,6 +8,12 @@ async function createApp(options: CreateAppOptions = {}) {
   const Koa = require('koa');
   const { ApolloServer } = require('../ApolloServer');
   const app = new Koa();
+  // Let's have errors be exposed to "users" instead of logged
+  // since it's a pain for us to check logs generically in this
+  // test suite.
+  app.on('error', (e: any) => {
+    e.expose = true;
+  });
 
   const server = new ApolloServer(
     (options.graphqlOptions as Config) || { schema: Schema },
@@ -21,12 +27,12 @@ async function destroyApp(app: any) {
   if (!app || !app.close) {
     return;
   }
-  await new Promise(resolve => app.close(resolve));
+  await new Promise((resolve) => app.close(resolve));
 }
 
 describe('koaApollo', () => {
   const { ApolloServer } = require('../ApolloServer');
-  it('throws error if called without schema', function() {
+  it('throws error if called without schema', function () {
     expect(() => new ApolloServer(undefined as any)).toThrow(
       'ApolloServer requires options.',
     );
@@ -34,5 +40,5 @@ describe('koaApollo', () => {
 });
 
 describe('integration:Koa', () => {
-  testSuite({createApp, destroyApp});
+  testSuite({ createApp, destroyApp });
 });